Redirect Only divert global package cache in Azure Pipelines
On local machines, file locks on the global package cache can cause `git clean -fdx` to fail. It also causes package restore times to be longer on local boxes after cleaning.
The extra protection afforded by a diverted package cache is good, but is probably sufficient to provide only in Azure Pipelines, or explicitly by the user when they want it.