Add MessagePackReader.ReadRaw()
We already had ReadRaw(long) where the caller had to know the number of bytes to skip. But if the caller just wants to skip a token (regardless of its length) then they had to call Skip(), which didn't return the token as well. To gain access to the sequence that was skipped, there was no public API to make that simple. ReadRaw() (no length argument) made the most sense to add since it's just like its preexisting ReadRaw(long) counterpart, but figures out the length itself.

Alternative considered:
1. Since this behaves more like Skip(), except that it returns the slice that was skipped, I considered simply changing the Skip() method to always return the slice. However since slicing has shown up as a perf bottleneck in some contexts, and skipping is supposed to be as fast as possible, I didn't want to add a mandatory slice to every skip operation that many callers would ignore.
2. Calling the new method SkipAndRead(). I think this suggests "skipping then reading" which is not what happens.