Ensure that only documented exceptions are thrown
We catch and wrap all exceptions thrown from the MessagePackSerializer APIs.

This ensures that arbitrary formatters or resolvers can't throw exceptions outside the documented set and cause unexpected failures in applications that are prepared to only catch the documented exception. The pattern of wrapping exceptions in outer exceptions is quite common and conducive to progressively precise errors that aid investigations.

In doing this, a few tests failed that were still expecting different exception types, so I've updated these tests. One in particular revealed that the DynamicObjectResolver was generating code that still threw thew old InvalidOperationException type so I fixed that as well.