Fix MessagePackSerializer.Deserialize(Stream) to fix-up stream position
This fixes the bug that the Stream deserializing methods would advance the stream position unless it was a `MemoryStream` in which case it left the position alone. That was an observable side effect of the `MemoryStream` optimization, which wasn't documented.

This also enhances the stream deserializing methods to not just advance them all alike but to advance them only to the end of the deserialized msgpack token. Since we don't know in advance how far to read, I kept the behavior of draining the stream, but then if the stream is seekable, we "rewind" the stream back to the position of the first byte that was not deserialized. This enables an intuitive use case as shown in the tests I added of writing several values to a stream then reading several values from the stream and it actually works.

Fixes #662